font-checker/requirements.txt
flask
python-pptx
olefile
hwp5
beautifulsoup4